pyproject.toml
[project]
name = "henhouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
    "ultralytics>=8.3.162",
    "requests>=2.25.0",
    "ha-mqtt-discoverable>=0.19.0",
    "torch>=2.6.0,<2.7.0",
    "torchvision>=0.21.0,<0.22.0",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.0",
    "psutil>=5.8.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
